refactor(precompile-macros): move contract logic out of lib.rs into modules
Extract supporting types and helpers from lib.rs into dedicated modules so
lib.rs contains only thin proc-macro entry points (Rust requires #[proc_macro]
functions to live in the crate root):

- contract.rs: ContractConfig, FieldInfo, FieldKind, parse_fields, gen_storage
- test_fields.rs: gen_test_fields_layout / gen_test_fields_struct helpers

lib.rs is now 80 lines of thin stubs that delegate to modules, consistent
with the workspace CLAUDE.md rule that lib.rs files must be minimal with no logic.
